@@ -14,7 +14,7 @@
     "better-react-mathjax": "^2.0.4-beta1",
     "lucide-react": "^0.469.0",
     "mongodb": "^6.13.0",
-    "next": "15.1.3",
+    "next": "^15.2.4",
     "next-themes": "^0.4.4",
     "react": "19.0.0",
     "react-dom": "19.0.0",
