@@ -2,13 +2,9 @@ import { MCQClient } from "@/components/mcq/MCQClient";
 import { MCQ } from "@/types/mcq";
 import { getApiUrl } from "@/utils/api";
 
-interface MCQPageProps {
-  params: { id: string };
-}
-
 async function getMCQData(id: string): Promise<MCQ> {
   try {
-    const url = await getApiUrl(`/api/data?key=${encodeURIComponent(id)}`);
+    const url = await getApiUrl(`/api/mcqs?key=${encodeURIComponent(id)}`);
 
     const res = await fetch(url, {
       cache: "no-store",
@@ -21,8 +17,7 @@ async function getMCQData(id: string): Promise<MCQ> {
       throw new Error(`Failed to fetch MCQ data (Status: ${res.status})`);
     }
 
-    const data = await res.json();
-    return data;
+    return res.json();
   } catch (error) {
     console.error("Error fetching MCQ:", error);
     throw error;
@@ -30,22 +25,12 @@ async function getMCQData(id: string): Promise<MCQ> {
 }
 
 // 🔹 Page Component
-export default async function MCQPage({ params }: MCQPageProps) {
+export default async function MCQPage({ params }: { params: { id: string } }) {
   try {
-    const resolvedParams = await params; // Ensure params is awaited
-    const mcqData = await getMCQData(resolvedParams.id);
-
+    const mcqData = await getMCQData(params.id);
     return <MCQClient initialMCQ={mcqData} />;
   } catch (error) {
-    console.error("Error loading MCQ:", error);
-
-    return (
-      <div className="flex items-center justify-center min-h-screen text-red-600">
-        ❌ Error loading MCQ. Please try again later.
-      </div>
-    );
+    console.log(error);
+    return <div>Error loading MCQ</div>;
   }
 }
-
-// 🔹 Dynamic Page Handling
-export const dynamic = "force-dynamic"; // Ensures fresh data on each request
