@@ -3,10 +3,6 @@ import { SubjectivePaper } from "@/types/subjective";
 import path from "path";
 import { promises as fs } from "fs";
 
-interface SubjectivePageProps {
-  params: { id: string };
-}
-
 // async function getSubjectiveData(id: string) {
 //   // Replace with your actual API endpoint
 //   const res = await fetch(
@@ -42,13 +38,21 @@ async function getSubjectiveData(id: string): Promise<SubjectivePaper> {
   return paper;
 }
 
-export default async function SubjectivePage({ params }: SubjectivePageProps) {
-  const resolvedParams = await params; // Ensure params is awaited
-  const data = await getSubjectiveData(resolvedParams.id);
+export default async function SubjectivePage({
+  params,
+}: {
+  params: { id: string };
+}) {
+  try {
+    const data = await getSubjectiveData(params.id);
 
-  return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      <SubjectiveClient data={data} />
-    </main>
-  );
+    return (
+      <main className="flex min-h-screen flex-col items-center justify-between p-24">
+        <SubjectiveClient data={data} />
+      </main>
+    );
+  } catch (error) {
+    console.log(error);
+    return <div>Error loading Subjective questions</div>;
+  }
 }
