@@ -34,7 +34,7 @@ const SubjectiveClient = ({ data }: SubjectiveClientProps) => {
               Section {section.sectionName}
             </h2>
             <p className="text-gray-600 mt-2">{section.description}</p>
-            {section.questions.map((question, index) => {
+            {section.questions.map((question) => {
               // console.log(index, question); // Logs each question before rendering
 
               return (
